--- src/query/planner/src/plans/create_database.rs
@@ -1,62 +1,0 @@
-// Copyright 2022 Datafuse Labs.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-use std::sync::Arc;
-
-use common_datavalues::DataSchema;
-use common_datavalues::DataSchemaRef;
-use common_meta_app::schema::CreateDatabaseReq;
-use common_meta_app::schema::DatabaseMeta;
-use common_meta_app::schema::DatabaseNameIdent;
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct CreateDatabasePlan {
-    pub if_not_exists: bool,
-    pub tenant: String,
-    pub catalog: String,
-    pub database: String,
-    pub meta: DatabaseMeta,
-}
-
-impl From<CreateDatabasePlan> for CreateDatabaseReq {
-    fn from(p: CreateDatabasePlan) -> Self {
-        CreateDatabaseReq {
-            if_not_exists: p.if_not_exists,
-            name_ident: DatabaseNameIdent {
-                tenant: p.tenant,
-                db_name: p.database,
-            },
-            meta: p.meta,
-        }
-    }
-}
-
-impl From<&CreateDatabasePlan> for CreateDatabaseReq {
-    fn from(p: &CreateDatabasePlan) -> Self {
-        CreateDatabaseReq {
-            if_not_exists: p.if_not_exists,
-            name_ident: DatabaseNameIdent {
-                tenant: p.tenant.clone(),
-                db_name: p.database.clone(),
-            },
-            meta: p.meta.clone(),
-        }
-    }
-}
-
-impl CreateDatabasePlan {
-    pub fn schema(&self) -> DataSchemaRef {
-        Arc::new(DataSchema::empty())
-    }
-}
