+++ src/query/planner/src/plans/mod.rs
@@ -17,7 +17,6 @@ mod alter_udf;
 mod alter_user;
 mod alter_view;
 mod call;
-mod create_database;
 mod create_role;
 mod create_stage;
 mod create_udf;
@@ -25,7 +24,6 @@ mod create_user;
 mod create_view;
 mod delete;
 mod describe_table;
-mod drop_database;
 mod drop_role;
 mod drop_stage;
 mod drop_table;
@@ -40,27 +38,22 @@ mod list;
 mod optimize_table;
 mod projection;
 mod remove_stage;
-mod rename_database;
 mod rename_table;
 mod revoke_privilege;
 mod revoke_role;
 mod set_role;
 mod setting;
-mod show_create_database;
 mod show_create_table;
 mod show_grants;
 mod show_roles;
 mod truncate_table;
-mod undrop_database;
 mod undrop_table;
-mod use_database;
 
 pub use alter_table_cluster_key::AlterTableClusterKeyPlan;
 pub use alter_udf::AlterUDFPlan;
 pub use alter_user::AlterUserPlan;
 pub use alter_view::AlterViewPlan;
 pub use call::CallPlan;
-pub use create_database::CreateDatabasePlan;
 pub use create_role::CreateRolePlan;
 pub use create_stage::CreateStagePlan;
 pub use create_udf::CreateUDFPlan;
@@ -68,7 +61,6 @@ pub use create_user::CreateUserPlan;
 pub use create_view::CreateViewPlan;
 pub use delete::*;
 pub use describe_table::DescribeTablePlan;
-pub use drop_database::DropDatabasePlan;
 pub use drop_role::DropRolePlan;
 pub use drop_stage::DropStagePlan;
 pub use drop_table::DropTablePlan;
@@ -84,19 +76,14 @@ pub use optimize_table::OptimizeTableAction;
 pub use optimize_table::OptimizeTablePlan;
 pub use projection::*;
 pub use remove_stage::RemoveStagePlan;
-pub use rename_database::RenameDatabaseEntity;
-pub use rename_database::RenameDatabasePlan;
 pub use rename_table::RenameTableEntity;
 pub use rename_table::RenameTablePlan;
 pub use revoke_privilege::RevokePrivilegePlan;
 pub use revoke_role::RevokeRolePlan;
 pub use set_role::SetRolePlan;
 pub use setting::*;
-pub use show_create_database::ShowCreateDatabasePlan;
 pub use show_create_table::ShowCreateTablePlan;
 pub use show_grants::ShowGrantsPlan;
 pub use show_roles::ShowRolesPlan;
 pub use truncate_table::TruncateTablePlan;
-pub use undrop_database::UndropDatabasePlan;
 pub use undrop_table::UndropTablePlan;
-pub use use_database::UseDatabasePlan;
