+++ src/query/service/src/interpreters/interpreter_database_create.rs
@@ -16,7 +16,7 @@ use std::sync::Arc;
 
 use common_exception::ErrorCode;
 use common_exception::Result;
-use common_planner::plans::CreateDatabasePlan;
+use common_sql::plans::CreateDatabasePlan;
 use common_users::UserApiProvider;
 
 use crate::interpreters::Interpreter;
