+++ src/query/service/src/interpreters/interpreter_database_drop.rs
@@ -15,7 +15,7 @@
 use std::sync::Arc;
 
 use common_exception::Result;
-use common_planner::plans::DropDatabasePlan;
+use common_sql::plans::DropDatabasePlan;
 
 use crate::interpreters::Interpreter;
 use crate::pipelines::PipelineBuildResult;
