+++ src/query/service/tests/it/storages/fuse/table_test_fixture.rs
@@ -25,8 +25,8 @@ use common_datavalues::prelude::*;
 use common_exception::Result;
 use common_meta_app::schema::DatabaseMeta;
 use common_planner::extras::Extras;
-use common_planner::plans::CreateDatabasePlan;
 use common_sql::executor::table_read_plan::ToReadDataSourcePlan;
+use common_sql::plans::CreateDatabasePlan;
 use common_storage::StorageFsConfig;
 use common_storage::StorageParams;
 use common_storages_fuse::FUSE_TBL_XOR_BLOOM_INDEX_PREFIX;
