+++ src/query/sql/src/planner/plans/plan.rs
@@ -26,7 +26,6 @@ use common_planner::plans::AlterUDFPlan;
 use common_planner::plans::AlterUserPlan;
 use common_planner::plans::AlterViewPlan;
 use common_planner::plans::CallPlan;
-use common_planner::plans::CreateDatabasePlan;
 use common_planner::plans::CreateRolePlan;
 use common_planner::plans::CreateStagePlan;
 use common_planner::plans::CreateUDFPlan;
@@ -34,7 +33,6 @@ use common_planner::plans::CreateUserPlan;
 use common_planner::plans::CreateViewPlan;
 use common_planner::plans::DeletePlan;
 use common_planner::plans::DescribeTablePlan;
-use common_planner::plans::DropDatabasePlan;
 use common_planner::plans::DropRolePlan;
 use common_planner::plans::DropStagePlan;
 use common_planner::plans::DropTableClusterKeyPlan;
@@ -48,20 +46,16 @@ use common_planner::plans::GrantRolePlan;
 use common_planner::plans::ListPlan;
 use common_planner::plans::OptimizeTablePlan;
 use common_planner::plans::RemoveStagePlan;
-use common_planner::plans::RenameDatabasePlan;
 use common_planner::plans::RenameTablePlan;
 use common_planner::plans::RevokePrivilegePlan;
 use common_planner::plans::RevokeRolePlan;
 use common_planner::plans::SetRolePlan;
 use common_planner::plans::SettingPlan;
-use common_planner::plans::ShowCreateDatabasePlan;
 use common_planner::plans::ShowCreateTablePlan;
 use common_planner::plans::ShowGrantsPlan;
 use common_planner::plans::ShowRolesPlan;
 use common_planner::plans::TruncateTablePlan;
-use common_planner::plans::UndropDatabasePlan;
 use common_planner::plans::UndropTablePlan;
-use common_planner::plans::UseDatabasePlan;
 
 use crate::optimizer::SExpr;
 use crate::plans::copy_v2::CopyPlanV2;
@@ -78,8 +72,14 @@ use crate::plans::share::RevokeShareObjectPlan;
 use crate::plans::share::ShowGrantTenantsOfSharePlan;
 use crate::plans::share::ShowObjectGrantPrivilegesPlan;
 use crate::plans::share::ShowSharesPlan;
+use crate::plans::CreateDatabasePlan;
+use crate::plans::DropDatabasePlan;
 use crate::plans::KillPlan;
+use crate::plans::RenameDatabasePlan;
+use crate::plans::ShowCreateDatabasePlan;
+use crate::plans::UndropDatabasePlan;
 use crate::plans::UpdatePlan;
+use crate::plans::UseDatabasePlan;
 use crate::BindContext;
 use crate::MetadataRef;
